pyproject.toml
[tool]
name = "CodeRefine"
version = "0.1.0"
description = "An AI tool that analyzes user input code, corrects errors, and provides suggestions for improvements and optimizations."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
radio = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"